AI-2922: move local_backend out of tools/ into its own top-level package
Rename src/keboola_mcp_server/tools/local/ → src/keboola_mcp_server/local_backend/
and tests/tools/local/ → tests/local_backend/.

This clearly separates the experimental local-backend from the main tools
package. The only change outside the local_backend directory is the import
in server.py (one line). All internal imports updated accordingly.

The local-backend is intentionally maintained at a lower bar than the main
codebase: it is experimental, Docker/local-only, and excluded from the
platform Keboola service.
