+++ src/keboola_mcp_server/local_backend/__init__.py
@@ -1,0 +1,4 @@
+from keboola_mcp_server.local_backend.backend import LocalBackend
+from keboola_mcp_server.local_backend.tools import register_local_tools
+
+__all__ = ['LocalBackend', 'register_local_tools']
