--- src/keboola_mcp_server/tools/local/__init__.py
@@ -1,4 +1,0 @@
-from keboola_mcp_server.tools.local.backend import LocalBackend
-from keboola_mcp_server.tools.local.tools import register_local_tools
-
-__all__ = ['LocalBackend', 'register_local_tools']
